requirements.txt
langchain>=0.3.25
langchain-core>=0.3.63
langchain-openai>=0.3.19
langchain-community>=0.3.21
mlrun==1.8.0
unstructured==0.16.12
markdown==3.7
langgraph>=0.4.8
gradio==4.44.1 # might need to be installed separately
numpy==1.26.4
s3fs==2024.6.1
duckduckgo-search==8.0.2
evaluate==0.4.3
transformers==4.48.1
torch==1.13.1
pydantic==2.10.6
fastapi==0.112.2
langchain-milvus==0.2.0